msm: camera: isp: Add SFE core idx in CSID error notify
It's possible that for a given use-case the pipeline is
CSID2->SFE0->IFE2, in the event of an overflow we need to log
the right SFE's debug registers. Unlike IFE we cannot use
CSID hw_idx to find the SFE core in the pipeline. The change
ensures we log the connected SFE core's debug registers.

CRs-Fixed: 2841729
Change-Id: I2beda7496b95a2b7c1aafc8cfcfcdde662c31898
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>