msm: camera: isp: Remove HW restriction on SFE RDI0-2 inputs
On Waipio SFE HPG recommends usage of RDI0-2 in sync with RD0-2.
RDI3-4 to be used for meta data. But with OEMs wanting to use 3 RDIs
for meta data for sHDR we need to break this constraint. This change
allows for SFE RDIs to be acquired independent of SFE RDs.
The change also configures scratch buffers accordingly.

CRs-Fixed: 2841729
Change-Id: I4fa84bb34c51f4cff40f931df0f80b259528b4da
Signed-off-by: Karthik Anantha Ram <kartanan@codeaurora.org>