qcacld-3.0: consider aggregated packets for packet drop
Currently, in the DP RX thread packet drop logic, the aggregated
packet as a whole is considered as one single packet. An aggregated
packet can hold anywhere between 2-16 packets, considering it as
a single packet will make the entire drop logic incorrect.

Consider all packets in the aggregate while computing the number of
enqueues & dequeues to fix the problem.

Change-Id: I70e65bcdfdbaf7821b36606ac6eeae61eaca7406
CRs-Fixed: 2948104
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>