qcacld-3.0: Don't set the hw_mode to SMM if connection is in progress
Currently, hdd_vdev_destroy checks if any vdev is actively using
DBS while cleaning up the vdev. It stops the timer and sends set
hw_mode command to firmware to set the hw_mode to SMM if no vdev
is using DBS.
This causes delay in p2p interface down in below scenario,
1. Bring up both sta and p2p vdevs.
2. Initiate connection on sta to an AP which has both 2GHz and
   5GHz candidates. hw_mode would be set to DBS as one of the
   candidate is 2GHz.
3. Connection would be attempted first on 5GHz candidate if the
   5GHz candidate score is good.
4. If p2p interface is disabled while sta connection is happening,
   it stops the opportunistic timer and sets the hw_mode to SMM.
   But the command would be waiting in serialization queue as
   connect command is active currently. This is a blocking call
   and userspace thread would be stuck till connection completion/
   connection timeout(5 seconds) happens.

The above wait results delay in p2p down which in turn results
delay in wifi turn-on(as p2p down+set random mac+p2p up happens
as part of wifi turn on) and turn-off.

Don't set the hw_mode to SMM in this case as the sta connection
is in progress.

Change-Id: If286fe9bd9906eba71686cf1d50880aa065cc100
CRs-Fixed: 2926272
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>