qcacld-3.0: Update WTC params only if the WTC attribute is present
Currently driver updates the WTC global parameters irrespective
of the WTC attributes present in the vendor command.
This causes the WTC parameters to be reset and WTC roam
trigger doesn't happen.
For example, with below sequence of commands:
wpa_cli DRIVER setwtcmode 0 1 -40 -95 -90 -85
wpa_cli DRIVER ADDROAMSCANFREQUENCIES_LEGACY 19 2412 2417 2422
The WTC parameters TLV is not included in the RSO
update command.

Update the WTC params only if the WTC attribute is present.

Change-Id: I53e5d934368169069dbf059335859936d320aa4e
CRs-Fixed: 2954458
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>