qcacld-3.0: Don't send DS param IE for 6G band
Currently, host sends DS param to fw based on
channel number but channel number may be same
in case of 6G and 2G due to which IEs can be
included in case of 6G which is not required.

Fix is to send DS param based on freq to
avoid addition of IEs in case of 6G band.

Change-Id: Ib711731fda3cfb0117478e6b1e026f40639a7fbb
CRs-Fixed: 2942323
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>