qcacld-3.0: SAP not following STA channel
With below INI configuration, SAP is not following the STA
channel. It's forming STA + SAP MCC instead of Forced SCC.
1. gWlanMccToSccSwitchMode = 4
   (QDF_MCC_TO_SCC_SWITCH_WITH_FAVORITE_CHANNE)
2. gEnableSAPManadatoryChanList=1

When channel switch mode is
QDF_MCC_TO_SCC_SWITCH_WITH_FAVORITE_CHANNE and if STA
is on 5ghz and SAP also comes on 5ghz then driver considers
the HW mode as DBS capable and it allows SAP to form
MCC which is not favorable.

For non_DBS target, the fix is to allow SAP to
do force SCC if the channel switch mode is
MCC_TO_SCC_SWITCH_WITH_FAVORITE_CHANNE.

Change-Id: I4b945dc1831909caff72f8398c959f4450765f2b
CRs-Fixed: 2972273
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>